requirements.txt
absl-py==2.3.1
accelerate==1.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.4.0
albucore==0.0.13
albumentations==1.4.10
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
astor==0.8.1
attrs==25.3.0
av==15.0.0
beautifulsoup4==4.13.4
bitsandbytes==0.45.3
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
contourpy==1.3.2
cycler==0.12.1
Cython==3.1.2
datasets==3.3.2
decorator==5.2.1
deepspeed==0.16.4
diffusers==0.33.1
dill==0.3.8
distro==1.9.0
docker-pycreds==0.4.0
einops==0.8.1
fairscale==0.4.13
fastapi==0.115.11
filelock==3.18.0
fire==0.7.0
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2024.12.0
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.44
h11==0.16.0
hf-xet==1.1.8
hf_transfer==0.1.9
hjson==3.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.29.1
idna==3.10
image-quality==1.2.7
imageio==2.37.0
imgaug==0.4.0
importlib_metadata==8.7.0
iopath==0.1.10
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
kiwisolver==1.4.8
lazy_loader==0.4
Levenshtein==0.27.1
libsvm==3.23.0.4
lmdb==1.7.2
lxml==6.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
ml_collections==1.1.0
modelscope==1.28.0
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.3
multiprocess==0.70.16
networkx==3.5
ninja==1.11.1.4
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
open_clip_torch==3.1.0
openai==1.97.0
opencv-contrib-python==4.11.0.86
opencv-python==4.8.0.76
opencv-python-headless==4.11.0.86
opt-einsum==3.3.0
packaging==25.0
paddleocr==2.9.1
paddlepaddle-gpu==2.6.2
pandas==2.2.3
peft==0.10.0
pillow==10.4.0
piq==0.8.0
platformdirs==4.3.8
portalocker==3.2.0
propcache==0.3.2
protobuf==5.29.5
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==21.0.0
pyclipper==1.3.0.post6
pycocotools==2.0.10
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-Levenshtein==0.27.1
pytorch-fid==0.3.0
pytz==2025.2
PyWavelets==1.9.0
PyYAML==6.0.2
RapidFuzz==3.13.0
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-image==0.21.0
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
sentencepiece==0.2.0
sentry-sdk==2.33.1
setproctitle==1.3.6
shapely==2.1.1
simsimd==6.5.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
starlette==0.46.2
stringzilla==3.12.5
sympy==1.13.1
termcolor==3.1.0
threadpoolctl==3.6.0
tifffile==2025.6.11
tiktoken==0.11.0
timm==1.0.19
tokenizers==0.19.1
tomli==2.2.1
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.40.0
triton==3.2.0
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.34.0
vendi-score==0.0.3
wandb==0.18.7
wcwidth==0.2.13
xformers==0.0.29.post3
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
